+++ cabby/rvs/speak_test.py
@@ -1,0 +1,30 @@
+# coding=utf-8
+# Copyright 2020 Google LLC
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+'''Tests for speak.py'''
+
+from cabby.rvs import speak
+
+import unittest
+
+class SpeakTest(unittest.TestCase):
+
+  def testSingleOutput(self):
+    expected = 'go to the pharmacy near the Empire State Building.'
+    output = speak.describe_route('Empire State Building', 'pharmacy')
+    self.assertEqual(output, expected)
+
+if __name__ == "__main__":
+    unittest.main()
+
